+++ src/packages/tabs/demo.taro.tsx
@@ -1,9 +1,31 @@
-import React, { useEffect, useRef, useState } from 'react'
+import React from 'react'
 import Taro from '@tarojs/taro'
-import { Star } from '@nutui/icons-react-taro'
-import { Tabs, Swiper } from '@/packages/nutui.react.taro'
 import { useTranslate } from '@/sites/assets/locale/taro'
 import Header from '@/sites/components/header'
+
+import Demo1 from './demos/taro/demo1'
+import Demo2 from './demos/taro/demo2'
+import Demo3 from './demos/taro/demo3'
+import Demo4 from './demos/taro/demo4'
+import Demo5 from './demos/taro/demo5'
+import Demo6 from './demos/taro/demo6'
+import Demo7 from './demos/taro/demo7'
+import Demo8 from './demos/taro/demo8'
+import Demo9 from './demos/taro/demo9'
+import Demo10 from './demos/taro/demo10'
+import Demo11 from './demos/taro/demo11'
+import Demo12 from './demos/taro/demo12'
+import Demo13 from './demos/taro/demo13'
+import Demo14 from './demos/taro/demo14'
+import Demo15 from './demos/taro/demo15'
+import Demo16 from './demos/taro/demo16'
+import Demo17 from './demos/taro/demo17'
+import Demo18 from './demos/taro/demo18'
+import Demo19 from './demos/taro/demo19'
+import Demo20 from './demos/taro/demo20'
+import Demo21 from './demos/taro/demo21'
+import Demo22 from './demos/taro/demo22'
+import Demo23 from './demos/taro/demo23'
 
 interface T {
   basic: string
@@ -13,29 +35,21 @@ interface T {
   titleButton: string
   titleDivider: string
   title2: string
+  title14: string
+  title10: string
+  title9: string
   title3: string
   title4: string
   title5: string
   title6: string
   title12: string
   title13: string
-  title14: string
   title7: string
   title8: string
-  title9: string
-  title10: string
   title11: string
   titleLeftCard: string
   titleLeftButton: string
   titleLeftDivider: string
-  custom1: string
-  custom2: string
-  custom3: string
-  pane1: string
-  pane2: string
-  pane3: string
-  pane4: string
-  pane5: string
 }
 
 const TabsDemo = () => {
@@ -63,14 +77,6 @@ const TabsDemo = () => {
       titleLeftCard: '左对齐-卡片模式',
       titleLeftButton: '左对齐-按钮模式',
       titleLeftDivider: '左对齐-分割线模式',
-      custom1: '自定义 1',
-      custom2: '自定义 2',
-      custom3: '自定义 3',
-      pane1: '低阶特卖',
-      pane2: '上新日',
-      pane3: '百亿补贴',
-      pane4: '今日聚超值',
-      pane5: '真好真便宜',
     },
     'en-US': {
       basic: 'Basic Usage',
@@ -95,485 +101,63 @@ const TabsDemo = () => {
       titleLeftCard: 'Title Left Align - Card Mode',
       titleLeftButton: 'Title Left Align - Button Mode',
       titleLeftDivider: 'Title Left Align - Divider Mode',
-      custom1: 'Custom 1',
-      custom2: 'Custom 2',
-      custom3: 'Custom 3',
-      pane1: 'Low-end Sale',
-      pane2: 'New Day',
-      pane3: 'Ten Billion Subsidies',
-      pane4: 'Super Value Today',
-      pane5: 'So Good And So Cheap',
     },
   })
-
-  const [tab1value, setTab1value] = useState<string | number>('0')
-  const [tab11value, setTab11value] = useState<string | number>('0')
-  const [tab111value, setTab111value] = useState<string | number>('0')
-  const [tab12value, setTab12value] = useState<string | number>('0')
-  const [tab2value, setTab2value] = useState<string | number>('0')
-  const [tab3value, setTab3value] = useState<string | number>('0')
-  const [tab4value, setTab4value] = useState<string | number>('0')
-  const [tab5value, setTab5value] = useState<string | number>('0')
-  const [tab51value, setTab51value] = useState<string | number>('0')
-  const [tab6value, setTab6value] = useState<string | number>('0')
-  const [tab61value, setTab61value] = useState<string | number>('0')
-  const [tab7value, setTab7value] = useState<string | number>('c1')
-  const [tab8value, setTab8value] = useState<string | number>('0')
-  const [tab81value, setTab81value] = useState<string | number>('0')
-  const [tab82value, setTab82value] = useState<string | number>('0')
-  const [tab9value, setTab9value] = useState<string | number>('0')
-  const [tab91value, setTab91value] = useState<string | number>('0')
-  const [tab92value, setTab92value] = useState<string | number>('0')
-  const [tab93value, setTab93value] = useState<string | number>('0')
-  const [list8, setList8] = useState<any>([])
-  const list4 = Array.from(new Array(10).keys())
-  const list5 = Array.from(new Array(2).keys())
-  const list6 = [
-    {
-      title: translated.custom1,
-      value: 'c1',
-      icon: <Star size={16} style={{ marginRight: '4px' }} />,
-    },
-    {
-      title: translated.custom2,
-      value: 'c2',
-    },
-    {
-      title: translated.custom3,
-      value: 'c3',
-    },
-  ]
-
-  useEffect(() => {
-    setTimeout(() => {
-      setTab8value(2)
-      setList8(Array.from(new Array(3).keys()))
-    }, 3000)
-  }, [])
-
-  const swiperRef = useRef<any>(null)
-  const [tabIndex, setTabIndex] = useState<string | number>(0)
 
   return (
     <>
       <Header />
       <div className={`demo ${Taro.getEnv() === 'WEB' ? 'web' : ''} full `}>
         <h2>{translated.basic}</h2>
-        <Tabs
-          value={tab1value}
-          onChange={(value) => {
-            setTab1value(value)
-          }}
-        >
-          <Tabs.TabPane title="Tab 1"> Tab 1</Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2"> Tab 2 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3"> Tab 3 </Tabs.TabPane>
-        </Tabs>
+        <Demo1 />
         <h2>{translated.title1}</h2>
-        <Tabs
-          value={tab2value}
-          onChange={(value) => {
-            setTab2value(value)
-          }}
-          activeType="smile"
-        >
-          <Tabs.TabPane title="Tab 1"> Tab 1 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2"> Tab 2 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3"> Tab 3 </Tabs.TabPane>
-        </Tabs>
+        <Demo2 />
         <h2>{translated.titleLite}</h2>
-        <Tabs
-          value={tab111value}
-          activeType="simple"
-          onChange={(value) => {
-            setTab111value(value)
-          }}
-        >
-          <Tabs.TabPane title="Tab 1"> Tab 1</Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2"> Tab 2 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3"> Tab 3 </Tabs.TabPane>
-        </Tabs>
+        <Demo3 />
         <h2>{translated.titleCard}</h2>
-        <Tabs
-          value={tab111value}
-          activeType="card"
-          onChange={(value) => {
-            setTab111value(value)
-          }}
-        >
-          <Tabs.TabPane title="Tab 1"> Tab 1</Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2"> Tab 2 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3"> Tab 3 </Tabs.TabPane>
-        </Tabs>
+        <Demo4 />
         <h2>{translated.titleButton}</h2>
-        <Tabs
-          value={tab111value}
-          activeType="button"
-          onChange={(value) => {
-            setTab111value(value)
-          }}
-        >
-          <Tabs.TabPane title="Tab 1"> Tab 1</Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2"> Tab 2 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3"> Tab 3 </Tabs.TabPane>
-        </Tabs>
+        <Demo5 />
         <h2>{translated.titleDivider}</h2>
-        <Tabs
-          value={tab111value}
-          activeType="divider"
-          onChange={(value) => {
-            setTab111value(value)
-          }}
-        >
-          <Tabs.TabPane title="Tab 1"> Tab 1</Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2"> Tab 2 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3"> Tab 3 </Tabs.TabPane>
-        </Tabs>
+        <Demo6 />
         <h2>{translated.title11}</h2>
-        <Tabs
-          value={tab3value}
-          align="left"
-          onChange={(value) => {
-            setTab3value(value)
-          }}
-        >
-          <Tabs.TabPane title="Tab 1"> Tab 1</Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2"> Tab 2 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3"> Tab 3 </Tabs.TabPane>
-        </Tabs>
+        <Demo7 />
         <h2>{translated.titleLeftCard}</h2>
-        <Tabs
-          value={tab111value}
-          activeType="card"
-          align="left"
-          onChange={(value) => {
-            setTab111value(value)
-          }}
-        >
-          <Tabs.TabPane title="Tab 1"> Tab 1</Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2"> Tab 2 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3"> Tab 3 </Tabs.TabPane>
-        </Tabs>
+        <Demo8 />
         <h2>{translated.titleLeftButton}</h2>
-        <Tabs
-          value={tab111value}
-          activeType="button"
-          align="left"
-          onChange={(value) => {
-            setTab111value(value)
-          }}
-        >
-          <Tabs.TabPane title="Tab 1"> Tab 1</Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2"> Tab 2 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3"> Tab 3 </Tabs.TabPane>
-        </Tabs>
+        <Demo9 />
         <h2>{translated.titleLeftDivider}</h2>
-        <Tabs
-          value={tab111value}
-          activeType="divider"
-          align="left"
-          onChange={(value) => {
-            setTab111value(value)
-          }}
-        >
-          <Tabs.TabPane title="Tab 1"> Tab 1</Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2"> Tab 2 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3"> Tab 3 </Tabs.TabPane>
-        </Tabs>
+        <Demo10 />
         <h2>{translated.title2}</h2>
-        <Tabs
-          value={tab4value}
-          onChange={(value) => {
-            setTab4value(value)
-          }}
-        >
-          <Tabs.TabPane title="Tab 1" value="0">
-            Tab 1
-          </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2" value="1" disabled>
-            Tab 2
-          </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3" value="2">
-            Tab 3
-          </Tabs.TabPane>
-        </Tabs>
+        <Demo11 />
         <h2>{translated.title14}</h2>
-        <Tabs
-          value={tabIndex}
-          onChange={(page) => {
-            swiperRef.current?.to(page)
-            setTabIndex(page)
-          }}
-        >
-          <Tabs.TabPane title="Tab 1" />
-          <Tabs.TabPane title="Tab 2" />
-          <Tabs.TabPane title="Tab 3" />
-        </Tabs>
-        <Swiper
-          defaultValue={0}
-          loop={false}
-          ref={swiperRef}
-          height={50}
-          onChange={(page) => {
-            setTabIndex(page.detail.current)
-          }}
-        >
-          <Swiper.Item>
-            <div style={{ backgroundColor: '#fff', padding: '10px' }}>
-              Tab 1
-            </div>
-          </Swiper.Item>
-          <Swiper.Item>
-            <div style={{ backgroundColor: '#fff', padding: '10px' }}>
-              Tab 2
-            </div>
-          </Swiper.Item>
-          <Swiper.Item>
-            <div style={{ backgroundColor: '#fff', padding: '10px' }}>
-              Tab 3
-            </div>
-          </Swiper.Item>
-        </Swiper>
+        <Demo12 />
         <h2>{translated.title10}</h2>
-        <Tabs
-          value={tab51value}
-          style={{ position: 'relative', zIndex: 11 }}
-          tabStyle={{ position: 'sticky', top: '0px', zIndex: 11 }}
-          onChange={(value) => {
-            setTab51value(value)
-          }}
-        >
-          <Tabs.TabPane title="Tab 1" value="0">
-            <p>Tab 1</p>
-            <p>Tab 1</p>
-            <p>Tab 1</p>
-            <p>Tab 1</p>
-            <p>Tab 1</p>
-            <p>Tab 1</p>
-            <p>Tab 1</p>
-            <p>Tab 1</p>
-            <p>Tab 1</p>
-          </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2" value="1">
-            <p>Tab 2</p>
-            <p>Tab 2</p>
-            <p>Tab 2</p>
-            <p>Tab 2</p>
-            <p>Tab 2</p>
-            <p>Tab 2</p>
-            <p>Tab 2</p>
-            <p>Tab 2</p>
-          </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3" value="2">
-            Tab 3
-          </Tabs.TabPane>
-        </Tabs>
+        <Demo13 />
         <h2>{translated.title9}</h2>
-        <Tabs
-          value={tab61value}
-          autoHeight
-          onChange={(value) => {
-            setTab61value(value)
-          }}
-        >
-          <Tabs.TabPane title="Tab 1" value="0">
-            <p>Tab 1</p>
-            <p>Tab 1</p>
-            <p>Tab 1</p>
-            <p>Tab 1</p>
-          </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2" value="1">
-            Tab 2
-          </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3" value="2">
-            Tab 3
-          </Tabs.TabPane>
-        </Tabs>
+        <Demo14 />
         <h2>{translated.title3}</h2>
-        <Tabs
-          value={tab81value}
-          onChange={(value) => {
-            setTab81value(value)
-          }}
-        >
-          {list8.map((item: any) => (
-            <Tabs.TabPane key={item} title={`Tab ${item}`}>
-              Tab {item}
-            </Tabs.TabPane>
-          ))}
-        </Tabs>
+        <Demo15 />
         <h2>{translated.title4}</h2>
-        <Tabs
-          name="a1"
-          value={tab91value}
-          onChange={(value) => {
-            setTab91value(value)
-          }}
-        >
-          <Tabs.TabPane title={translated.pane1}>
-            {translated.pane1}
-          </Tabs.TabPane>
-          <Tabs.TabPane title={translated.pane2}>
-            {translated.pane2}
-          </Tabs.TabPane>
-          <Tabs.TabPane title={translated.pane3}>
-            {translated.pane3}
-          </Tabs.TabPane>
-          <Tabs.TabPane title={translated.pane4}>
-            {translated.pane4}
-          </Tabs.TabPane>
-          <Tabs.TabPane title={translated.pane5}>
-            {translated.pane5}
-          </Tabs.TabPane>
-        </Tabs>
+        <Demo16 />
         <h2>{translated.title4} 2</h2>
-        <Tabs
-          name="a2"
-          value={tab92value}
-          style={{ height: '300px' }}
-          onChange={(value) => {
-            setTab92value(value)
-          }}
-          direction="vertical"
-        >
-          {list4.map((item) => (
-            <Tabs.TabPane key={item} title={`Tab ${item}`}>
-              Tab {item}
-            </Tabs.TabPane>
-          ))}
-        </Tabs>
+        <Demo17 />
         <h2>{translated.title5}</h2>
-        <Tabs
-          style={{ height: '300px' }}
-          value={tab5value}
-          onChange={(value) => {
-            setTab5value(value)
-          }}
-          direction="vertical"
-        >
-          {list5.map((item) => (
-            <Tabs.TabPane key={item} title={`Tab ${item}`}>
-              Tab {item}
-            </Tabs.TabPane>
-          ))}
-        </Tabs>
+        <Demo18 />
         <h2>{translated.title6}</h2>
-        <Tabs
-          style={{ height: '300px' }}
-          value={tab6value}
-          onChange={(value) => {
-            setTab6value(value)
-          }}
-          activeType="smile"
-          direction="vertical"
-        >
-          {list5.map((item) => (
-            <Tabs.TabPane key={item} title={`Tab ${item}`}>
-              Tab {item}
-            </Tabs.TabPane>
-          ))}
-        </Tabs>
+        <Demo19 />
 
         <h2>{translated.title12}</h2>
-
-        <Tabs
-          value={tab82value}
-          onChange={(value) => {
-            setTab82value(value)
-          }}
-          direction="vertical"
-        >
-          <Tabs.TabPane title="Tab 1">
-            <Tabs
-              value={tab9value}
-              onChange={(value) => {
-                setTab9value(value)
-              }}
-              direction="horizontal"
-            >
-              <Tabs.TabPane title="Tab 1"> Tab 1 </Tabs.TabPane>
-              <Tabs.TabPane title="Tab 2"> Tab 2 </Tabs.TabPane>
-              <Tabs.TabPane title="Tab 3"> Tab 3 </Tabs.TabPane>
-            </Tabs>
-          </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2"> Tab 2 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3"> Tab 3 </Tabs.TabPane>
-        </Tabs>
+        <Demo20 />
 
         <h2>{translated.title13}</h2>
-        <Tabs
-          value={tab8value}
-          onChange={(value) => {
-            setTab8value(value)
-          }}
-          autoHeight
-        >
-          <Tabs.TabPane title="Tab 1">
-            <Tabs
-              value={tab93value}
-              onChange={(value) => {
-                setTab93value(value)
-              }}
-              direction="vertical"
-            >
-              <Tabs.TabPane title="Tab 1"> Tab 1 </Tabs.TabPane>
-              <Tabs.TabPane title="Tab 2"> Tab 2 </Tabs.TabPane>
-              <Tabs.TabPane title="Tab 3"> Tab 3 </Tabs.TabPane>
-            </Tabs>
-          </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2"> Tab 2 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3"> Tab 3 </Tabs.TabPane>
-        </Tabs>
+        <Demo21 />
 
         <h2>{translated.title7}</h2>
-        <Tabs
-          value={tab11value}
-          onChange={(value) => {
-            setTab11value(value)
-          }}
-          style={{ '--nutui-tabs-titles-font-size': '20px' }}
-        >
-          <Tabs.TabPane title="Tab 1"> Tab 1 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2"> Tab 2 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3"> Tab 3 </Tabs.TabPane>
-        </Tabs>
-        <Tabs
-          value={tab12value}
-          onChange={(value) => {
-            setTab12value(value)
-          }}
-          style={{ '--nutui-tabs-titles-font-size': '12px' }}
-        >
-          <Tabs.TabPane title="Tab 1"> Tab 1 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 2"> Tab 2 </Tabs.TabPane>
-          <Tabs.TabPane title="Tab 3"> Tab 3 </Tabs.TabPane>
-        </Tabs>
+        <Demo22 />
 
         <h2>{translated.title8}</h2>
-        <Tabs
-          value={tab7value}
-          title={() => {
-            return list6.map((item) => (
-              <div
-                onClick={() => setTab7value(item.value)}
-                className={`nut-tabs-titles-item ${
-                  tab7value === item.value ? 'nut-tabs-titles-item-active' : ''
-                }`}
-                key={item.value}
-              >
-                {item.icon || null}
-                <span className="nut-tabs-titles-item-text">{item.title}</span>
-                <span className="nut-tabs-titles-item-line" />
-              </div>
-            ))
-          }}
-        >
-          {list6.map((item) => (
-            <Tabs.TabPane key={item.value} value={item.value}>
-              {item.title}
-            </Tabs.TabPane>
-          ))}
-        </Tabs>
+        <Demo23 />
       </div>
     </>
   )
