+++ src/packages/tabs/demos/taro/demo11.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import { Tabs } from '@nutui/nutui-react-taro'
 
 const Demo11 = () => {
-  const [tab2value, setTab2value] = useState('0')
+  const [tab2value, setTab2value] = useState<string | number>('0')
   return (
     <>
       <Tabs
@@ -12,16 +12,13 @@ const Demo11 = () => {
         }}
       >
         <Tabs.TabPane title="Tab 1" value="0">
-          {' '}
-          Tab 1{' '}
+          Tab 1
         </Tabs.TabPane>
         <Tabs.TabPane title="Tab 2" value="1" disabled>
-          {' '}
-          Tab 2{' '}
+          Tab 2
         </Tabs.TabPane>
         <Tabs.TabPane title="Tab 3" value="2">
-          {' '}
-          Tab 3{' '}
+          Tab 3
         </Tabs.TabPane>
       </Tabs>
     </>
