+++ src/packages/tabs/demos/taro/demo15.tsx
@@ -2,12 +2,12 @@ import React, { useState, useEffect } from 'react'
 import { Tabs } from '@nutui/nutui-react-taro'
 
 const Demo15 = () => {
-  const [tab3value, setTab3value] = useState(0)
-  const [list3, setList3] = useState()
+  const [tab3value, setTab3value] = useState<string | number>(0)
+  const [list3, setList3] = useState<number[]>([])
   useEffect(() => {
     setTimeout(() => {
       setTab3value(2)
-      setList3(Array.from(new Array(3).keys()))
+      setList3([...new Array(3).keys()])
     }, 3000)
   }, [])
   return (
@@ -20,8 +20,7 @@ const Demo15 = () => {
       >
         {list3.map((item) => (
           <Tabs.TabPane key={item} title={`Tab ${item}`}>
-            {' '}
-            Tab {item}{' '}
+            Tab {item}
           </Tabs.TabPane>
         ))}
       </Tabs>
