+++ src/packages/tabs/demos/taro/demo19.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import { Tabs } from '@nutui/nutui-react-taro'
 
 const Demo19 = () => {
-  const [tab6value, setTab6value] = useState('0')
+  const [tab6value, setTab6value] = useState<string | number>('0')
   const list5 = Array.from(new Array(2).keys())
   return (
     <>
@@ -17,8 +17,7 @@ const Demo19 = () => {
       >
         {list5.map((item) => (
           <Tabs.TabPane key={item} title={`Tab ${item}`}>
-            {' '}
-            Tab {item}{' '}
+            Tab {item}
           </Tabs.TabPane>
         ))}
       </Tabs>
